Pierwsze wersja backend, przygotwoanie o implementacji frontend. Brak lokalnego AI. Poprawnie dzialanie modelu rodziny, dodanie [Authoriztaion] do kontrolerów

+++ src/Backend/Controllers/ExaminationController.cs
@@ -2,11 +2,14 @@
 using Microsoft.EntityFrameworkCore;
 using Backend.Data;
 using Backend.DTOs.Examination;
+using Backend.Helpers;
 using Backend.Models;
 using Backend.Services;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Backend.Controllers;
 
+[Authorize]
 [ApiController]
 [Route("api/[controller]")]
 public class ExaminationController : ControllerBase
@@ -14,20 +17,38 @@ public class ExaminationController : ControllerBase
     
     private readonly AppDbContext _context;
     private readonly IGoogleCalendarService _googleCalendarService;
-
-    public ExaminationController(AppDbContext context, IGoogleCalendarService googleCalendarService)
+    private readonly IFamilyAccessService _familyAccessService;
+    
+    public ExaminationController(AppDbContext context, IGoogleCalendarService googleCalendarService, IFamilyAccessService familyAccessService)
     {
         _context = context;
         _googleCalendarService = googleCalendarService;
+        _familyAccessService = familyAccessService;
     }
 
     [HttpPost] // CreateExamination
     public async Task<ActionResult<ResponseExaminationDto>> PostExamination(CreateExaminationDto dto)
     {
+        
+        var userId = this.GetUserId();
+        var userRole = this.GetUserRole();
+        
+        if (userRole == RoleUser.Child)
+        {
+            return Forbid();
+        }
+        
+        var targetUserId = dto.ForUserId ?? userId; // pomaga nam to określić dla kogo tworzymy badanie
+        
+        if (targetUserId != userId && !await _familyAccessService.IsParentOfChildAsync(userId, dto.ForUserId.Value))
+        {
+            return Forbid();   
+        }
+        
         var examination = new Examination
         {
             Id = Guid.NewGuid(),
-            UserId = dto.UserId,
+            UserId = targetUserId,
             Date = dto.Date,
             Time = dto.Time,
             Name = dto.Name,
@@ -53,7 +74,13 @@ public class ExaminationController : ControllerBase
     [HttpGet] // ResponseExamination
     public async Task<ActionResult<IEnumerable<ResponseExaminationDto>>> GetExaminations()
     {
-        var examinations = await _context.Examinations.ToListAsync();
+        var userId = this.GetUserId();
+        var visibleUserIds = await _familyAccessService
+            .GetVisibleUserIdsAsync(userId);
+        var examinations = await _context.Examinations
+            .Where(e => visibleUserIds.Contains(e.UserId))
+            .Where(e => !_context.ExaminationsHide.Any(h => h.ExaminationId == e.Id && h.HiddenForUserId == userId))
+            .ToListAsync();
 
         return Ok(examinations.Select(ToResponseDto));
     }
@@ -61,7 +88,12 @@ public class ExaminationController : ControllerBase
     [HttpGet("{id}")]
     public async Task<ActionResult<ResponseExaminationDto>> GetExamination(Guid id)
     {
-        var examination = await _context.Examinations.FindAsync(id);
+        var userId = this.GetUserId();
+        var visibleUserIds = await _familyAccessService.GetVisibleUserIdsAsync(userId);
+        var examination = await _context.Examinations
+            .Where(e => e.Id == id && visibleUserIds.Contains(e.UserId))
+            .Where(e => !_context.ExaminationsHide.Any(h => h.ExaminationId == e.Id && h.HiddenForUserId == userId))
+            .FirstOrDefaultAsync();
         if (examination == null)
         {
             return NotFound($"Nie znaleziono badania o id {id}");
@@ -73,11 +105,29 @@ public class ExaminationController : ControllerBase
     [HttpPut("{id}")] // UpdateExaminationDto
     public async Task<ActionResult<ResponseExaminationDto>> PutExamination(Guid id, UpdateExaminationDto dto)
     {
-        var examination = await _context.Examinations.FindAsync(id);
-        if (examination == null)
+        var userId = this.GetUserId();
+
+        if (this.GetUserRole() == RoleUser.Child)
+        {
+            return Forbid();
+        }
+
+        var examination = await _context.Examinations
+            .FirstOrDefaultAsync(e => e.Id == id );
+        
+         if (examination == null)
         {
             return NotFound($"Nie znaleziono badania o id {id}");
         }
+         
+        var isOwner = examination.UserId == userId;
+        var isParent = await _familyAccessService.IsParentOfChildAsync(userId, examination.UserId);
+
+        if (!isOwner && !isParent)
+        {
+            return NotFound("Nie znaleziono badania");
+        }
+
 
         examination.Name = dto.Name;
         examination.Date = dto.Date;
@@ -105,8 +155,25 @@ public class ExaminationController : ControllerBase
     [HttpPatch("{id}")] // PatchExamination
     public async Task<ActionResult<ResponseExaminationDto>> PatchExamination(Guid id, PatchExaminationDto dto)
     {
-        var examination = await _context.Examinations.FindAsync(id);
+        var userId = this.GetUserId();
+
+        if (this.GetUserRole() == RoleUser.Child)
+        {
+            return Forbid();
+        }
+
+        var examination = await _context.Examinations
+            .FirstOrDefaultAsync(e => e.Id == id);
+        
         if (examination == null)
+        {
+            return NotFound($"Nie znaleziono badania o id {id}");
+        }
+        
+        var isOwner = examination.UserId == userId;
+        var isParent = await _familyAccessService.IsParentOfChildAsync(userId, examination.UserId);
+
+        if (!isOwner && !isParent)
         {
             return NotFound($"Nie znaleziono badania o id {id}");
         }
@@ -134,12 +201,29 @@ public class ExaminationController : ControllerBase
     [HttpDelete("{id}")] // DeleteExamination
     public async Task<IActionResult> DeleteExamination(Guid id)
     {
-        var examination = await _context.Examinations.FindAsync(id);
+        var userId = this.GetUserId();
+
+        if (this.GetUserRole() == RoleUser.Child)
+        {
+            return Forbid();
+        }
+
+        var examination = await _context.Examinations
+            .FirstOrDefaultAsync(e => e.Id == id);
+        
         if (examination == null)
         {
             return NotFound($"Nie znaleziono badania o id {id}");
         }
 
+        var isOwner = examination.UserId == userId;
+        var isParent = await _familyAccessService.IsParentOfChildAsync(userId, examination.UserId);
+
+        if (!isOwner && !isParent)
+        {
+            return NotFound($"Nie znaleziono badania o id {id}");
+        }
+        
         await _googleCalendarService.DeleteEventAsyncExamination(examination);
         
         _context.Examinations.Remove(examination);
@@ -149,6 +233,79 @@ public class ExaminationController : ControllerBase
         return NoContent();
     }
 
+    [HttpPost("{id}/hide/{hiddenForUserId}")]
+    public async Task<IActionResult> HideExamination(Guid id, Guid hiddenForUserId)
+    {
+        var userId = this.GetUserId();
+
+        if (this.GetUserRole() == RoleUser.Child)
+        {
+            return Forbid();
+        }
+
+        var examination = await _context.Examinations.FirstOrDefaultAsync(e => e.Id == id && e.UserId == userId);
+        if (examination == null)
+        {
+            return NotFound($"Nie znaleziono badania o id {id}");
+        }
+
+        if (hiddenForUserId == userId)
+        {
+            return BadRequest("Nie można ukryć badania przed samym sobą");
+        }
+        
+        var visibleUserIds = await _familyAccessService.GetVisibleUserIdsAsync(userId);
+        if (!visibleUserIds.Contains(hiddenForUserId))
+        {
+            return BadRequest("Ta osoba nie jest w twojej rodzinie");
+        }
+        
+        var aleardyHidden = await _context.ExaminationsHide
+            .AnyAsync(h => h.ExaminationId == id && h.HiddenForUserId == hiddenForUserId);
+        if (aleardyHidden)
+        {
+            return BadRequest("Baddanie jest ukryte");
+        }
+
+        _context.ExaminationsHide.Add(new ExaminationHide
+        {
+            Id = Guid.NewGuid(),
+            ExaminationId = id,
+            HiddenForUserId = hiddenForUserId,
+            HiddenByUserId = userId,
+            CreatedAt = DateTime.UtcNow,
+        });
+        
+        await _context.SaveChangesAsync();
+        
+        return  (Ok(ToResponseDto(examination)));
+    }
+
+    [HttpDelete("{id}/hide/{hiddenForUserId}")]
+    public async Task<IActionResult> UnhideExamination(Guid id, Guid hiddenForUserId)
+    {
+        var userId = this.GetUserId();
+        
+        var examination = await _context.Examinations
+            .FirstOrDefaultAsync(e => e.Id == id && e.UserId == userId);
+        if (examination == null)
+        {
+            return NotFound($"Nie znaleziono badania");
+        }
+        
+        var hide = await _context.ExaminationsHide
+            .FirstOrDefaultAsync(e => e.ExaminationId == id && e.HiddenForUserId == hiddenForUserId);
+        if (hide == null)
+        {
+            return NotFound($"Nie znaleziono badania");
+        }
+        
+        _context.ExaminationsHide.Remove(hide);
+        await _context.SaveChangesAsync();
+        
+        return Ok(ToResponseDto(examination));
+    }
+    
     private static ResponseExaminationDto ToResponseDto(Examination e)
     {
         return new ResponseExaminationDto
